server/requirements.txt
aioredis==2.0.1
fastapi==0.110.0
uvicorn>=0.27.1
python-multipart==0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]
sqlalchemy
fastapi-sso>=0.7.0
beanie>=1.22.6
mongomock>=4.1.2
mongomock-motor>=0.0.21
motor>=3.3.1
pymongo>=4.5.0
pydantic>=2.4.2
pydantic_core>=2.10.1
pydantic-settings
PyJWT

pytest
pytest-mock
pytest-asyncio
pytest-cov
httpx
